+++ spotbugs-tests/src/test/java/edu/umd/cs/findbugs/util/MutableClassesTest.java
@@ -91,6 +91,14 @@ class MutableClassesTest {
             return new Immutable(n);
         }
 
+        public void setNUnsupported(int n) {
+            throw new UnsupportedOperationException();
+        }
+
+        public void setNUnsupported2(int n) {
+            throw new UnsupportedOperationException("This class is immutable, setters are unsupported.");
+        }
+
         public static Immutable getImmutable() {
             return immutable;
         }